+++ internal/src/repo/mod.rs
@@ -1,2 +1,2 @@
 pub mod api;
-pub mod blog;
+pub mod blogs;
